Restructure imports.
