+++ src/common/index.ts
@@ -1,0 +1,11 @@
+/**
+ * index.ts
+ *
+ * Description:
+ *    Index file for shared logic and data.
+ *
+ */
+
+export { reactRoutes, resolveRoute } from './routes';
+export { getStatusCode } from './status-codes';
+export { createMessage, IPCRequestObject, IPCResponseObject } from './util';
