+++ src/renderer/Components/Auth/index.tsx
@@ -1,0 +1,13 @@
+/**
+ * index.tsx
+ *
+ * Description:
+ *    Index File for all Auth Components.
+ *
+ */
+
+import Login from './Login';
+import Register from './Register';
+
+// Re-exporting default exports as named exports.
+export { Login, Register };
