+++ src/renderer/Components/Elements/Message/index.tsx
@@ -7,8 +7,8 @@
  */
 
 import React, { useEffect } from 'react';
-import { useAppDispatch, useAppSelector } from '../../Hooks';
-import { clearMessageState } from '../../State/reducer';
+import { useAppDispatch, useAppSelector } from 'renderer/Hooks';
+import { clearMessageState } from 'renderer/State';
 import './message.scss';
 
 const Message = () => {
