+++ src/renderer/Components/Header/Navbar.tsx
@@ -6,11 +6,11 @@
  *
  */
 
+import IMAGE from 'assets/logo/png/256x256.png';
+import { reactRoutes } from 'common/routes';
 import React from 'react';
 import { Link } from 'react-router-dom';
-import IMAGE from '../../../assets/logo/png/256x256.png';
-import { reactRoutes } from '../../../common/routes';
-import { useAppSelector } from '../../Hooks';
+import { useAppSelector } from 'renderer/Hooks';
 import Logout from './Logout';
 
 const Navbar = () => {
