+++ src/renderer/Components/index.tsx
@@ -1,0 +1,17 @@
+/**
+ * index.tsx
+ *
+ * Description:
+ *    Index File for all Components.
+ *
+ */
+
+import Header from './Header';
+import Notes from './Notes';
+import Sidebar from './Sidebar';
+import Spaces from './Spaces';
+
+export { Login, Register } from './Auth';
+export { Button, Form, Input, Message } from './Elements';
+// Re-exporting default exports as named exports.
+export { Header, Notes, Sidebar, Spaces };
