+++ src/renderer/index.tsx
@@ -11,7 +11,7 @@ import { render } from 'react-dom';
 import { Provider } from 'react-redux';
 import { MemoryRouter as Router } from 'react-router-dom';
 import App from './App';
-import store from './State';
+import { store } from './State';
 import './style.scss';
 
 render(
